% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDensityPlot2.R
\name{scDensityPlot2}
\alias{scDensityPlot2}
\title{scDensityPlot function}
\usage{
scDensityPlot2(
  object = NULL,
  reduction = "umap",
  title = NULL,
  split.by = NULL,
  ncol = NULL,
  colors = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction to use, default reduction="umap"}

\item{title}{Figure title}

\item{split.by}{Name of a metadata column to split plot by}

\item{ncol}{Number of columns for display the plots}

\item{colors}{Vector of colours to use for n-colour gradient}
}
\value{
ggplot object
}
\description{
Visualize density plot in low dimensional embedding
}
\author{
Wei Zhang
}
