% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDGE.R
\name{scDGE}
\alias{scDGE}
\title{Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters}
\usage{
scDGE(
  object = object,
  comparison = c("condA", "condB"),
  group.by = NULL,
  min.cells = 20,
  logFC = 0.25,
  clusters = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{comparison}{Vector of nominal variables from group.by. Eg., comparison=c('Normal', 'Tumor')}

\item{group.by}{Regroup cells before performing DGE, default group.by='group'}

\item{min.cells}{Minimum number of cells to perform DGE, default min.cells=20}

\item{logFC}{A positive value to set the cutoff of logFC, default logFC=0.25}

\item{clusters}{Vector of Clusters to perform DGE, default clusters=NULL,
which means all clusters will be performed}
}
\value{
DGE Table
}
\description{
Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters
}
\author{
rstatistics
}
