% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDimPlot.R
\name{scDimPlot}
\alias{scDimPlot}
\title{scDimPlot}
\usage{
scDimPlot(
  object = NULL,
  cols = NULL,
  pt.size = NULL,
  reduction = NULL,
  split.by = NULL,
  label = TRUE,
  title = NULL,
  ncol = NULL,
  raster = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{cols}{Colors to plot, use ggplot2's default colors by default. We include a pallete called 'sc' which consists of 36 colors}

\item{pt.size}{Adjust point size to plot, default pt.size=0.5}

\item{reduction}{Which dimensionality reduction to use}

\item{split.by}{Name of a metadata column to split plot by}

\item{label}{Whether to label the clusters}

\item{title}{Title of the plot}

\item{ncol}{Number of columns for display the plots}

\item{raster}{Convert points to raster format, default is TRUE}
}
\value{
ggplot2 object
}
\description{
DimPlot with rasterized point for single-cell visualization
}
