% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFeatureDotPlot.R
\name{scFeatureDotPlot}
\alias{scFeatureDotPlot}
\title{scFeatureDotPlot}
\usage{
scFeatureDotPlot(
  object,
  features,
  cols = NULL,
  dot.scale = NULL,
  font.size = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of features to plot}

\item{cols}{Vector of two colors to form the gradient over}

\item{dot.scale}{Scale the size of the dot}

\item{font.size}{Font size}
}
\description{
scFeatureDotPlot
wrapper for Seurat::DotPlot
}
\author{
Wei Zhang
}
