% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFeaturePlot.R
\name{scFeaturePlot}
\alias{scFeaturePlot}
\title{scScoreDimPlot}
\usage{
scFeaturePlot(
  object = NULL,
  features = NULL,
  reduction = NULL,
  cols = NULL,
  pt.size = NULL,
  split.by = NULL,
  title = NULL,
  ncol = NULL,
  raster = TRUE,
  scale = TRUE,
  col.min = NA,
  col.max = NA
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Name of gene symbols}

\item{reduction}{Which dimensionality reduction to use}

\item{cols}{Colors to plot}

\item{pt.size}{Adjust point size to plot, default pt.size=0.5}

\item{split.by}{Name of a metadata column to split plot by}

\item{title}{Title of the plot}

\item{ncol}{Number of columns for display the plots}

\item{raster}{Convert points to raster format, default is TRUE}

\item{scale}{Determine whether to scale the data, default is TRUE}

\item{col.min}{Minimum scaled average score threshold (smaller values will be set to this)}

\item{col.max}{Maximum scaled average score threshold (larger values will be set to this)}
}
\value{
ggplot2 object
}
\description{
MSigdb scoring DimPlot for single-cell clusters
}
