% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFeaturePlot2.R
\name{scFeaturePlot2}
\alias{scFeaturePlot2}
\title{scFeaturePlot}
\usage{
scFeaturePlot2(
  object = object,
  reduction = NULL,
  features = features,
  pt.size = NULL,
  font.size = NULL,
  colors = NULL,
  ncol = NULL,
  combine = FALSE,
  order = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction to use}

\item{features}{Vector of features to plot}

\item{pt.size}{Point size to plot}

\item{font.size}{Font size}

\item{colors}{Vector of two colors to form the gradient over}

\item{ncol}{Number of columns to combine features}

\item{combine}{Combine plots into a single patchworked ggplot object}

\item{order}{Determine whether to plot cells in order of expression}
}
\description{
scFeaturePlot
wrapper for Seurat::FeaturePlot
}
\author{
rstatistics
}
