% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHeatmap.R
\name{scHeatmap}
\alias{scHeatmap}
\title{scHeatmap visualization}
\usage{
scHeatmap(
  object = NULL,
  key = NULL,
  cluster = NULL,
  feature.by = NULL,
  markers = NULL,
  colors = NULL,
  top_n = 5,
  font.size = 8,
  assay = NULL,
  slot = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The key slot that stored the DGE table}

\item{cluster}{Cluster specified by the user}

\item{feature.by}{Feature which is used to divid cells into two groups}

\item{markers}{A vector of features to plot, default is top_n features}

\item{colors}{User defined colors}

\item{top_n}{Top n markers to plot}

\item{font.size}{Font size}

\item{assay}{Assay to pull from}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}
}
\description{
Average Heatmap for specific feature in all clusters
}
\author{
rstatistics
}
