% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegrate.R
\name{scIntegrate}
\alias{scIntegrate}
\title{scIntegrate}
\usage{
scIntegrate(
  object.list = NULL,
  object.names = NULL,
  nVariable = 2000,
  nPC = 30,
  res = 0.5,
  batch.rm = "harmony"
)
}
\arguments{
\item{object.list}{A list of seurat objects}

\item{object.names}{An array of seurat object names}

\item{nVariable}{Number of features to select as top variable features}

\item{nPC}{Number of PCs to use}

\item{res}{Resolution parameter to set. Default res=0.5}

\item{batch.rm}{Remove batch effect with 'seurat' or 'harmony'. 
Default batch.rm='harmony'}
}
\value{
Seurat object.
}
\description{
Integrate seurat objects into one
}
\author{
rstatistics
}
