% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMSigdbScoring.R
\name{scMsigdbScoring}
\alias{scMsigdbScoring}
\title{scMsigdbScoring}
\usage{
scMsigdbScoring(object = NULL, category = NULL, geneSets = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{category}{String of specific MSigDB category to use, 
default category='H' stands for 'hallmark gene sets'}

\item{geneSets}{Vector of gene sets, default geneSets='NULL' 
stands for all gene sets in category will be used}
}
\value{
Seurat object
}
\description{
MSigdb signature scoring for single-cell clusters
}
\author{
Wei Zhang
}
