% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPathogenDGE.R
\name{scPathogenDGE}
\alias{scPathogenDGE}
\title{Find differentially expressed genes by species for each clusters in a dataset}
\usage{
scPathogenDGE(
  object = NULL,
  species.by = NULL,
  clusters = NULL,
  min.cells = 20
)
}
\arguments{
\item{object}{Seurat object}

\item{species.by}{A vector of features which is used to divid cells into two groups}

\item{clusters}{Clusters selected to do DGE}

\item{min.cells}{Minimum number of cells to do DGE}
}
\value{
DGE table
}
\description{
Find differentially expressed genes by species for each clusters in a dataset
}
\author{
Wei Zhang
}
