% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scScoreDotPlot.R
\name{scScoreDotPlot}
\alias{scScoreDotPlot}
\title{scScoreDotPlot}
\usage{
scScoreDotPlot(
  object = NULL,
  geneSets = NULL,
  cols = c("blue", "orange", "red"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  group.by = NULL,
  split.by = NULL,
  scale = TRUE,
  scale.min = NA,
  scale.max = NA
)
}
\arguments{
\item{object}{Seurat object}

\item{geneSets}{Vector of gene sets to display}

\item{cols}{Colors to plot, default cols=c('blue', 'orange', 'red')}

\item{col.min}{Minimum scaled average score threshold}

\item{col.max}{Maximum scaled average score threshold}

\item{dot.min}{Lower limit of cell fraction to draw the dots (default is 0)}

\item{dot.scale}{Scale the size of the dots}

\item{group.by}{Group the cells by}

\item{split.by}{Split the cells by}

\item{scale}{Whether to scale the data, default scale = TRUE}

\item{scale.min}{Lower limit to scale the data, default scale.min = NA}

\item{scale.max}{Upper limit to scale the data, default scale.max = NA}
}
\description{
MSigdb Scoring DotPlot for single-cell clusters
}
\author{
rstatistics
}
