% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scOne.R
\name{scOne}
\alias{scOne}
\title{single-cell analysis of one sample}
\usage{
scOne(
  object = NULL,
  nFeatures = 2000,
  nPC = 30,
  resolution = 0.5,
  perplexity = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{nFeatures}{Number of variable features to use}

\item{nPC}{Number of principal components to use}

\item{resolution}{Value of the resolution parameter}

\item{perplexity}{Value of the perplexity parameter, set between 5 and 50}
}
\value{
Seurat object
}
\description{
A wrapper of the standard workflow of Seurat
}
\author{
Wei Zhang
}
